requirements.txt
asgiref==3.6.0
autopep8==2.0.1
Django==4.1.5
django-ckeditor==6.5.1
django-js-asset==2.0.0
django-tinymce==3.5.0
lxml==4.9.2
Pillow==9.4.0
pycodestyle==2.10.0
python-docx==0.8.11
python-dotenv==1.0.0
reportlab==3.6.12
sqlparse==0.4.3
tomli==2.0.1
